requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
cssselect==1.1.0
cycler==0.10.0
idna==2.8
kiwisolver==1.1.0
lxml==4.4.1
matplotlib==3.1.1
mpl-finance==0.10.0
multidict==4.5.2
numpy==1.17.1
pandas==0.25.1
pyparsing==2.4.2
pyquery==1.4.0
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
six==1.12.0
slackclient==2.1.0
urllib3==1.25.3
yarl==1.3.0
